+++ packages/expo-contacts/CHANGELOG.md
@@ -17,6 +17,8 @@
 
 ### 💡 Others
 
+- Handle new permission status on `iOS` 18.
+
 ## 13.0.4 - 2024-06-10
 
 _This version does not introduce any user-facing changes._